add PIO_TYPENAME, rearrange code
