+++ utils/python/CIME/XML/env_mach_pes.py
@@ -18,91 +18,22 @@ class EnvMachPes(EnvBase):
         self._component_value_list = ["NTASKS", "NTHRDS", "NINST", "ROOTPE", "PSTRID"]
         self._components = []
 
-    def set_components(self, components):
-        if 'DRV' in components:
-            index = components.index("DRV")
-            components[index] = "CPL"
-        self._components = components
-
-    def check_if_comp_var(self, vid, attribute=None):
-        '''
-        Returns the varid and component for
-        '''
-        comp = None
-        if vid in self._component_value_list:
-            if attribute is not None:
-                if "component" in attribute:
-                    comp = attribute["component"]
-            return vid, comp, True
+    def comp_var_split(self, vid):
         parts = vid.split("_")
         if len(parts) == 2 and parts[0] in self._component_value_list:
-            iscompvar = True
             return parts[0], parts[1], True
         return vid, None, False
 
-
-
     def get_value(self, vid, attribute=None, resolved=True, subgroup=None, pes_per_node=None): # pylint: disable=arguments-differ
-        value = None
-        vid, comp, iscompvar = self.check_if_comp_var(vid, attribute)
-
-        if iscompvar and comp is None:
-            logger.debug("Not enough info to get value for %s"%vid)
-            return value
-        elif comp is not None:
-            if attribute is None:
-                attribute = {"component" : comp}
-            else:
-                attribute["component"] = comp
-            node = self.get_optional_node("entry", {"id":vid})
-            if node is not None:
-                type_str = self._get_type_info(node)
-                value = convert_to_type(self.get_element_text("value", attribute, root=node), type_str, vid)
-
-        if value is None:
-            value = EnvBase.get_value(self, vid, attribute, resolved, subgroup)
+        value = EnvBase.get_value(self, vid, attribute, resolved, subgroup)
 
         if "NTASKS" in vid or "ROOTPE" in vid and pes_per_node is None:
             pes_per_node = self.get_value("PES_PER_NODE")
-
             if "NTASKS" in vid and value < 0:
                 value = -1*value*pes_per_node
             if "ROOTPE" in vid and value < 0:
                 value = -1*value*pes_per_node
         return value
-
-    def set_value(self, vid, value, subgroup=None, ignore_type=False):
-        """
-        Set the value of an entry-id field to value
-        Returns the value or None if not found
-        subgroup is ignored in the general routine and applied in specific methods
-        """
-        vid, comp, iscompvar = self.check_if_comp_var(vid, None)
-        val = None
-        node = self.get_optional_node("entry", {"id":vid})
-        if node is not None:
-            if iscompvar and comp is None:
-                for comp in self._components:
-                    val = self._set_value(node, value, vid, subgroup, ignore_type, component=comp)
-            else:
-                val = self._set_value(node, value, vid, subgroup, ignore_type, component=comp)
-
-        return val
-
-    def _set_value(self, node, value, vid=None, subgroup=None, ignore_type=False, component=None): # pylint: disable=arguments-differ
-        if vid is None:
-            vid = node.get("id")
-
-        if vid in self._component_value_list:
-            attribute = {"component":component}
-            type_str = self._get_type_info(node)
-            # special case - no NINST defined for coupler component
-            if vid == "NINST" and component == "CPL":
-                return None
-            val = self.set_element_text("value", convert_to_string(value, type_str, vid), attribute, root=node)
-            return val
-        val = EnvBase._set_value(self, node, value, vid, subgroup, ignore_type)
-        return val
 
     def get_max_thread_count(self, comp_classes):
         ''' Find the maximum number of openmp threads for any component in the case '''
