+++ utils/python/CIME/XML/env_run.py
@@ -14,3 +14,14 @@ class EnvRun(EnvBase):
         initialize an object interface to file env_run.xml in the case directory
         """
         EnvBase.__init__(self, case_root, infile)
+        self._components = []
+        self._component_value_list = ["PIO_TYPENAME"]
+
+    def comp_var_split(self, vid):
+        parts = vid.split("_", 1)
+        if len(parts) == 2 and parts[1] in self._component_value_list:
+            return parts[1], parts[0], True
+        return vid, None, False
+
+
+
